[#104174950] switching to ewok
in switching to ewok we needed to change the box we
used.

this caused us to restructure and refactor a bit
in terms of our pipeline. pulling build back
into the build step where it should be,
and deploy only deploys...

this will also, likely fix the 1st failure bug #104117294.
we shall see