[#112467509] first broad stroke at plugin support
we now have the ability to create a plugin,
register the plugin with a cfops.

there are toggles powered by env vars
to turn this on and off, as well as
adjust the plugin dir target.

the pluginTileWrapper which will wrap
a plugin call in a tile interface is
not yet complete and cfops is not capable
of calling plugins.

it is not yet importing plugins to the tile
registry.... more to come

this includes samples and example plugins